Move video utils into `softlearning/utils/video.py`
