+++ examples/development/simulate_policy.py
@@ -9,7 +9,7 @@ import tensorflow as tf
 from softlearning.environments.utils import get_environment_from_params
 from softlearning.policies.utils import get_policy_from_variant
 from softlearning.samplers import rollouts
-from softlearning.misc.utils import save_video
+from softlearning.utils.video import save_video
 
 
 DEFAULT_RENDER_KWARGS = {
