+++ softlearning/algorithms/rl_algorithm.py
@@ -11,7 +11,7 @@ import tensorflow as tf
 from tensorflow.python.training import training_util
 
 from softlearning.samplers import rollouts
-from softlearning.misc.utils import save_video
+from softlearning.utils.video import save_video
 
 
 if LooseVersion(tf.__version__) > LooseVersion("2.00"):
